+++ sdk/python/inkbox/tunnels/__init__.py
@@ -7,6 +7,7 @@ from inkbox.tunnels.types import (
     TLSMode,
     Tunnel,
     TunnelStatus,
+    TunnelSummary,
 )
 from inkbox.tunnels.exceptions import (
     TunnelCSRStateConflict,
@@ -30,6 +31,7 @@ __all__ = [
     "TLSMode",
     "Tunnel",
     "TunnelStatus",
+    "TunnelSummary",
     "TunnelsResource",
     "TunnelCSRStateConflict",
     "TunnelError",